* Brought imap branch to PM land.

Conflicts:
	headers/os/app/Message.h
	src/add-ons/mail_daemon/inbound_filters/spam_filter/SpamFilter.cpp
	src/add-ons/mail_daemon/inbound_filters/spam_filter/SpamFilterConfig.cpp
	src/add-ons/mail_daemon/inbound_protocols/imap/Jamfile
	src/add-ons/mail_daemon/inbound_protocols/imap/imap_lib/IMAPProtocol.h
	src/kits/app/Message.cpp
	src/kits/mail/ProtocolConfigView.cpp
	src/kits/network/libnetapi/SecureSocket.cpp
	src/preferences/mail/ConfigWindow.cpp
	src/servers/mail/Jamfile

+++ src/tests/kits/interface/bregion/RegionTestcase.cpp
@@ -10,6 +10,8 @@
 #include "RegionTestcase.h"
 #include <Region.h>
 #include <Rect.h>
+
+#include <assert.h>
 #include <math.h>
 
 
@@ -106,7 +108,7 @@ RegionTestcase::RegionTestcase(std::string name) :
 RegionTestcase::~RegionTestcase()
 {
 	while(!listOfRegions.IsEmpty()) {
-		delete static_cast<BRegion *>(listOfRegions.RemoveItem((long int)0));
+		delete static_cast<BRegion *>(listOfRegions.RemoveItem(int32(0)));
 	}
 }
